mcp-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-firestore==2.19.0
google-api-python-client==2.149.0
google-auth==2.35.0
mcp==1.2.0
